requirements.txt
slackeventsapi
slackclient
